+++ src/components/login/index.jsx
@@ -1,19 +1,25 @@
-import React, { useState } from 'react'
-import { useDispatch } from 'react-redux'
-import { Link } from 'react-router-dom'
-import Wrapper from './style'
+import React, { useState } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
+import Wrapper from './style';
+import axios from 'axios';
 
 const Login = () => {
-  const [email, setEmail] = useState('')
-  const [password, setPasswod] = useState('')
-  const dispatch = useDispatch()
+  const [contact, setContact] = useState('');
+  const [password, setPassword] = useState(''); // Corrected typo here
+  const navigate = useNavigate();
+
   const login = (e) => {
-    e.preventDefault()
-    dispatch({
-      type: "LOG_IN",
-      payload: { email, password }
-    })
-  }
+    e.preventDefault();
+    axios.get('https://server-api1-li2k.onrender.com/api/user/login')
+      .then((res) => {
+        console.log(res.dat);
+        navigate('/profile');
+      })
+      .catch((error) => {  // Changed from .then to .catch for error handling
+        console.log('API Request Failed', error);
+      });
+  };
+
   return (
     <Wrapper>
       <div className="inner">
@@ -21,15 +27,15 @@ const Login = () => {
         <p>Welcome back!</p>
         <input
           type='text'
-          value={email}
-          placeholder='Email'
-          onChange={(e) => setEmail(e.target.value)}
+          value={contact}
+          placeholder='Contact'
+          onChange={(e) => setContact(e.target.value)}
         />
         <input
           type='password'
           value={password}
           placeholder='Password'
-          onChange={(e) => setPasswod(e.target.value)}
+          onChange={(e) => setPassword(e.target.value)} {/* Corrected the function name */}
         />
         <input
           type='button'
@@ -39,7 +45,7 @@ const Login = () => {
         <p className='bottom_text'>Don't have an account?<Link className='link' to='/signupform'>Sign Up</Link></p>
       </div>
     </Wrapper>
-  )
-}
+  );
+};
 
-export default Login
+export default Login;
